requirement.txt
Package                 Version
----------------------- -------
Click                   7.0    
Django                  2.1    
django-markdown-deux    1.0.5  
django-rq               1.3.0  
djangorestframework     3.9.0  
djangorestframework-jwt 1.11.0 
elasticsearch           6.4.0  
elasticsearch-dsl       6.2.1  
ipaddress               1.0.22 
LEPL                    5.1.3  
Markdown                3.0.1  
markdown2               2.3.7  
pip                     18.1   
PyJWT                   1.7.1  
python-dateutil         2.8.0  
pytz                    2018.7 
redis                   3.0.1  
rq                      0.13.0 
setuptools              40.6.3 
six                     1.12.0 
urllib3                 1.25.3 
wheel                   0.32.3 
